Make into package
Add GitHub Actions workflow for build release, add assembly definitions

https://github.com/koorimizuw/YamaPlayer/blob/222a7f63b1c47e50d733955a68de2cefc7396565/.github/workflows/release.yml
