+++ Documentation/ReadMe(Recorder).md
@@ -1,0 +1,7 @@
+# 導入の手順(Recorder)
+
+VCCのVRCSDK3-WorldとUdonSharpをAddしたのちに、RecorderのunitypackageをImportしてください。
+Assets\HUMR\Prefabs\Recorder.prefabをワールドSceneのHierarchyに設置してください。
+そのワールドをアップロード(またはLocalTest)してワールドに入りログを残します。
+VRChatを終了するとC:\Users\username\AppData\LocalLow\VRChat\VRChat\の下にoutput_log_xx-xx-xx.txtが作成されます。
+これによってOutputLogLoaderを利用する準備が整いました。
