+++ Runtime/CSharp/HumrUtilities.cs
