+++ Runtime/mukaderabbit.humr.Runtime.asmdef
@@ -1,0 +1,19 @@
+{
+    "name": "mukaderabbit.humr.Runtime",
+    "rootNamespace": "HUMR",
+    "references": [
+        "GUID:99835874ee819da44948776e0df4ff1d",
+        "GUID:77fc2e1f5a4f24efab94081ad67fcb20",
+        "GUID:0f9e8f73c69964f1e8176f9a0b86f9b2",
+        "GUID:84265b35cca3905448e623ef3903f0ff"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
