+++ Runtime/mukaderabbit.humr.Runtime.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: d1bf393c3e46e554d963a2c0d6d4df74
-folderAsset: yes
-DefaultImporter:
+guid: 4156b7325025f45479ad99b85e7efd7c
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
