+++ Runtime/mukaderabbit.humr.Runtime.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 78b965b503bad07419b69b59ca9b0330
-folderAsset: yes
-DefaultImporter:
+guid: 78dfdc72d7058de4a919359ccb0db652
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
