+++ TODO.md
@@ -1,3 +1,24 @@
-﻿# TODO
+# TODO
 
 - [ ] English and Japanese localization
+
+## Recorder.cs
+
+- [ ] Add begin and end tags to logs with data structures
+- [ ] Merge `InteractRecorder.cs`
+
+## Recorder.prefab
+
+- [ ] Add user interface with record button
+- [ ] T-pose on recording start
+
+## OutputLogLoader.cs
+
+- [ ] Make compatible with previous Log syntax
+- [ ] Fix hip position so feet are on the ground
+
+## OutputLogLoaderEditor.cs
+
+- [ ] Choose DisplayName from dropdown
+- [ ] Only show log files with HUMR data
+- [ ] Export Mode (Humanoid/Generic) dropdown
