+++ package.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fb34b25f33c6f214f9e52e2b31cbd15d
+PackageManifestImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
